Increment and jump
https://imcs.dvfu.ru/cats/static/problem_text-cpid-1092004.html

+++ inc_and_jmp.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 char a;
-int add_cnt = 0, result = 0;
+long long add_cnt = 0, result = 0;
 
 int main()
 {
